requirements.txt
dnspython>=2.4
